We had nothing. Here is something.
The code is a bit of a mess right now. I was trying different solutions and they merged into a monster.

The idea is to create a deck list with cards and values, shuffle the list, pull the first item in the list, pop that item so it can't be pulled again. Right now it doesn't work but my brain is fried for tonight so here is what I have right now.

I haven't added the scoring code yet.